Updated openstack/openstack Project: openstack/swift  bd4cb4490f02f6d2957a877ffdc2aa3b7ee07d1a
Skip SLO-reconciling probe test when SLO is off

The probe test in question tries to make a manifest referencing
unavailable objects and expects that to fail with a 400. If the SLO
middleware is enabled, then it checks the segments, can't find some,
and returns the 400. If it's disabled, the PUT succeeds and makes an
object whose contents are some JSON blob. In the latter case, the
probe test would fail because it expected a real SLO manifest but
didn't find one.

Now we skip the remainder of the test when we detect that SLO is not
enabled.

Change-Id: I3e7e8e98107608e675efc24156e703bc167458bb