Updated openstack/openstack Project: openstack/nova  755c2cb0c76a4c642b2d89053b63a528ee79013c
virt: reserved number of mempages on compute host

Users need to mark as reserved some amount of pages for third party
components.

The most common use case for using huge/large pages is NFV. In the
current state of that feature we can't guarantee the necessary amount
of pages to allow OVS-DPDK to run properly on the compute node, which
result in the instance failing to boot on a well selected
compute-node. OVS-DPDK needs 1 GB hugepages reserved. Since Nova does
not take into account that page reserved for OVS-DPDK it results in
the process not being able to acquire the necessary memory which
results in a failed boot.

This commit adds a new option 'reserved_huge_pages' which takes a list
of string format to select on which host NUMA nodes and from which
pagesize we want to reserve a certain amount of pages. It also updates
NUMAPageTopology to contain a reserved memory pages attribute, which
helps compute the available pages size on host for scheduling/claiming
resources.

Change-Id: Ie04d6362a4e99dcb2504698fc831a366ba746b44
Closes-Bug: #1543149