Updated openstack/openstack Project: openstack/keystone  64925bf698cb45fde3c0941cfcbec5b45d31a0d7
switch to tempest instead of deprecated tempest-lib

tempest-lib is deprecated and recommends us to use tempest instead.[1]

tempest actually is not required from our repository, because when
we run tempest, it must be installed and setup firstly (at least for
now). But when we generate document for our tempest test code, it is
required, so add it to test-requirements.txt.

[1]: https://github.com/openstack/tempest-lib

Change-Id: Id04f5340c27cfc84f4aba4732684a73a977ccec8