Updated openstack/openstack Project: openstack/nova  74b115a473b034781a51c7810ef6abbe3a7baed8
api-ref: parameter validation for os-security-group-rules

Cleans up a bunch of the types and descriptions for the
parameters.

Removes some extraneous unused parameters.

Added descriptions for the request and response to point out
that group_id and cidr are optional but one or the other is
required and depending on what comes in the request influences
what comes back in the response.

Part of blueprint api-ref-in-rst

Change-Id: I1084fa7ec46373cc989d1d93bb667ac59a62e38a