Updated openstack/openstack Project: openstack/cliff  0488b26777ba0dd71c9d94f771162fed2ac5000d
Fix prettytable.PrettyTable().max_width wrong usage

prettytable.PrettyTable().max_width sets *per-column* max_width
which is not what we want here. What we want to set is *total
line length*. So go through the _assign_max_widths() code that
does that for the other cases, and get it right.

Change-Id: I11b0b2ec18de8f59759f52647afba6e4ae7917c6
Signed-off-by: Vincent Legoll <vincent.legoll@idgrilles.fr>