Updated openstack/openstack Project: openstack/keystone  4fa6b08ea1034e55b98a4bdd3fbb5e0529ed311f
Move the federation abstract base class out of core

This patch moves the federation abstract base class out of core and into
backends/base.py

This removes dependencies where backend code references code in the
core. The reasoning being that the core should know about the backend
interface, but the backends should not know anything about the core
(separation of concerns). And part of the risk here is a potential for
circular dependencies.

Partial-Bug: #1563101

Change-Id: I3304f9c1fafe42cfe955f44224703ace82ae185a