Updated openstack/openstack Project: openstack/cinder  c4aa40a1670d2e575ca0c02ec5434072bfd77d6b
Cleanup orphaned code from image and manager

There were several methods in the image module and manager code that
were no longer being used anywhere. While some of these have
potential for being useful, they are not in use and should not be
sitting around the codebase waiting for someone to need them.

Change-Id: Id3f3320c4153dabad309699698b0e52f8312c081