Updated openstack/openstack Project: openstack/keystone  3600014b66856666a2612513fa623f68cb7583c0
Don't set None for ldap.OPT_X_TLS_CACERTFILE

The logic in the cleanups for these tests checks for
values of None which are not accepted by ldap.set_option().
In my local testing environment, possibly because I only run a small
subset of Keystone tests, the existing value for this option
is None, causing all tests to fail with
"TypeError: set_option() argument must be string, not None" [1].

Ideally, the function used to call ldap.set_option() should guard
against any values of None in all cases, however this is an immediate
illustration / fix for the exact issue upstream SQLAlchemy is seeing.

[1] https://jenkins.sqlalchemy.org/job/openstack/74/sqla_branch=rel_1_1/console

Change-Id: I1e65e8278c77c5c15f4be5b186ededa8297f9216