Updated openstack/openstack Project: openstack/neutron  080703d62c63025b20756cf294de524b67684e39
Added PecanResourceExtension

Extensions that implement the get_pecan_controllers should now return a list of
PecanResourceExtension.  This data structure is being added in place of a tuple
of tuples so that it's easier to know what attributes should be populated.
This is in preparation of allowing the addition of more values other than
collection and controller, like path_prefix.

Change-Id: I12937c02843725e0cad5f646526dc3bdcc51c30b