Updated openstack/openstack Project: openstack/requirements  658a6d2dee7c49ada949766c6b628b87bbe312b4
Revert removal of upper constraint on python-kafka

The python-kafka library at 1.0.0 is incompatible with prior
releases and introduced a significant number of performance
regressions and bugs. The Monasca project is a heavy user of
the python-kafka library and this change has broken it.
Unfortunately, we cant' move to the latest version of the
library until all the performance issues and bugs have been
resolved and verified.

We would like to revert back to the previous commit which had
upper contraints specified.

Note, we are trying to get all the Monasca libraries into
requirements, but that work hasn't been completed yet.

Change-Id: I78cf1c725afe7fb8942ff20f3d4332fad0ace8fa