Updated openstack/openstack Project: openstack/heat  ff00644f8f9e24fa9fd0b134206f73f54e706e14
Remove usages of :linenos: in docs

Some markup features are sphinx-specific and
are not supported by current library. Some bug were
reported https://bugs.launchpad.net/doc8/+bug/1426503
about this, but it is not fixed and we need working
docs.
In current situation whole blocks of code are not shown
so i removed usages of error-inflicting directive :linenos:

Change-Id: I25e2d193b10d713bbe64fb227afb71310874bb9f