Updated openstack/openstack Project: openstack/diskimage-builder  d253a94187d86940d9ee1364eaf910fd510d1704
Do not remove sudo in ironic-agent

"ironic-agent" element is currently removing sudo, which breaks other
elements such as devuser.  There appears to be no security or other
reason to do this, it's just the way it has always been.  Leave sudo
in as it is considered part of the base cloud images.

Change-Id: Ida9b1885f745146071e4b2d85ae59341ac85d5c8
Closes-Bug: #1572486