Updated openstack/openstack Project: openstack/diskimage-builder  e7625214f6523a276e8af27d02dfc931564b7fd8
Properly fail/trap in eval_run_d

In phases which are called from eval_run_d (block-device.d) we do not
listen to exit 1's nor do we allow break=after-error. This is because
the run_d function is called in a subshell in order to grab its output.

This also turns on pipefail in the main disk-image-create script.

Change-Id: I88ab2e7104148437eabfe6880e3a1e5ebbb2c15d