Updated openstack/openstack Project: openstack-dev/devstack  7cd7f0576bea4c1623dc8f1e217f422f49f6bcd7
Fix return code detection in plugin list generation

As can be seen in logs of the periodic generation job, our cgit does a
weird thing where sometimes it returns a 404 page with content, and
sometimes a zero response (see [1] for example, the last number is
response size).  This appears to be an openstack CI issue; possibly
due to cgit caching or similar (see [2] for manual test).  It will
have to be investigated with the host apache logs.

This is resulting in a lot of projects incorrectly being picked up as
having plugins (I7116571d2a2b1fc3a61e5f1ed46ac2cbc244775a).  I'm not
sure if this problem is also releated to the original status-code
issues mentioned in the code, but testing shows that cgit is correctly
returning 404's for missing files (you can see in the logs [1]).  Thus
switch the logic to examine the return code which avoids this issue.

[1] http://logs.openstack.org/periodic/propose-devstack-plugins-list/e55790c/console.html.gz#_2016-05-04_06_46_51_660
[2] http://paste.openstack.org/show/496434/

Change-Id: I6a06347d91d091441f6f7b70f99aba6d8e9add4b