Updated openstack/openstack Project: openstack/heat  151e9cfd84a09116c695b1b57dba9b9769de8350
Use ResourceDefinition API in resource_facade

Don't treat a ResourceDefinition as a snippet of CloudFormation template in
the resource_facade intrinsic function.

This is slightly tricky, because in the case of the update_policy we don't
necessarily have a schema and we just want to pass the (resolved) data
verbatim, so we have to access a non-public attribute of the object.

Change-Id: Ib23a4be4f0a250a292e0ec707b1d5e885f15ed7e