Updated openstack/openstack Project: openstack/oslo.messaging  3158d483f44ff49b15c42fca4fe13e27258a8480
Refactor RPC client

The RPCClient class is fully based on the _CallContext class.
I think it is redundant to have the same attributes in the RPCClient
as in the _CallContext just in order to pass instances with these
attributes to the _CallContext._prepare method (which is not so obious
because it can work with both _CallContext and RPCCLient classes,
though these classes are not in the same class hierarchy).

This patch tries to solve all these issues by including both classes
in a hierarchy and to make a little optimization by eliminating
dummy _CallContext instances creation on each cast/call in RPCClient.

Change-Id: I608c828123b196099933d16801f8ce7d9dd3a079