Updated openstack/openstack Project: openstack-infra/release-tools  2c805391ad17452d9ecb0ec179fa67f8f1c65d24
automatically submit constraint updates after tagging

After we tag a release, automatically propose the constraint
update. This is a stop-gap until we have the release automation going
further, at which point we should move this to its own job to run after
a package has been uploaded to pypi successfully (giving the constraints
tests a better chance of passing).

Change-Id: I1d2118dc93ec08bc2cfcd4b9dec0b30705ab5198
Signed-off-by: Doug Hellmann <doug@doughellmann.com>