Updated openstack/openstack Project: openstack/heat  fe6313cde3b59cb0617ce4018a0a818b72a8b81f
Reset env.encrypted_param_names on stack update

If encryption was enabled when the stack was created but then disabled
when the stack was updated, env.params and env.encrypted_param_names
will be in an inconsistent state.  Ensure that
env.encrypted_param_names is an empty list when encryption is
disabled.

Change-Id: Ife84348bd6b45e1174b3faef72b0b11f3bab2a3f
Closes-Bug: #1572294