Updated openstack/openstack Project: openstack/requirements  10db24c9f8d23eb3e6dfbbdc3d4f996ae981ea9a
Add Mistral to global requirements

This is for use within tripleo-common[0]. Version 2015.1.0 is excluded as pip
detects it as the latest version since mistral changed it's versioning
convention. The current latest version is 2.0.0, which was released over a
year later.

We can remove this when Mistral reaches a major version of 2016!

[0]: https://review.openstack.org/282366

Change-Id: Ibb98013d6ef283bd454f958bb036557d2b02997f