Updated openstack/openstack Project: openstack-dev/devstack  51c93b5780bc069a92cc568561f6ab59d0840b42
Add a deprecation warning for lib/neutron-legacy

This commit adds a deprecation warning for lib/neutron-legacy. Right
now lib/neutron isn't quite in a place where we can use it by default
but we're getting close. As soon as it's passing in the gate we plan
to make a switch over and a hard delete of lib/neutron-legacy. To give
any users which have a hard dependency on it (which is not actually
a supported use case) a heads up this adds the deprecation warning
in front of that change.

Change-Id: Idf1faf2e9dd497f9b97abfcc6e796ca72d60d955