Updated openstack/openstack Project: openstack/neutron  46716c88739907d30c21009d29bc0534e5a4b81e
Pecan: skip all notifications on failures

If there was a bad status code, none of the notification
logic should attempt to process. A bad status code means
that the assumptions the notification code makes about
original resources being present may not be true.

This also fixes the exception translator so the correct
status code is set by the hook on exceptions that have
a fault map entry.

Closes-Bug: #1557516
Change-Id: Ibff419e7caa6de422f04eb1e43fd2aed304bdd97