Updated openstack/openstack Project: openstack/nova  52a831e78e3bfbbaa1e129ed931a2d2b0dc48071
Switch api unit tests to use v2.1 API

TestNeutronSecurityGroupsOutputTest and BootFromVolumeTest were tests
only for legacy v2 API code, and v2.1 API were not tested on the same
test cases. This patch makes the tests work for v2.1 API code for the
test coverage.

In addition, this patch changes a request body of BootFromVolumeTest
because v2.1 API validation detected the bugs of the body like the
following:

* Invalid input for field/attribute volume_id. Value: 1. u'1' is not a 'uuid'"
* Additional properties are not allowed (u'virtual' was unexpected)

Change-Id: I9c2a53679ed6a097b062ba542dddd85a42c9964f