Updated openstack/openstack Project: openstack/nova  76fd3d013d52225d4f96c4cde40333a688ac47ba
Add a note about egress rules to os-security-group-rules api-ref

There have been at least a couple of bugs about not being able
to create egress security group rules in Nova, which is because
nova-network does not support them. Neutron does, but Nova does
not proxy this to Neutron, nor will it.

So add a note in the api-ref docs for creating security group
rules about the egress rule limitation with nova-network.

Change-Id: Idc79cd1718b52db8611fd108b23f176f925221a6
Related-Bug: #1579749
Related-Bug: #1267140