Updated openstack/openstack Project: openstack/requirements  2e7df9c0acf460bdc0dd2a29bb9d388287dab98f
add a validation command to ensure projects use check-requirements

Add a command and tox environment to let us define a job to ensure that
all repos in projects.txt have the check-requirements job defined. Two
projects are missing the job now, so this patch depends on the patch to
add those jobs.

Change-Id: I61658a7265a77e72401978dda558f41b28a81d92
Depends-On: I81e1c70257e28583e776171580d75341e620bf8d
Signed-off-by: Doug Hellmann <doug@doughellmann.com>