Updated openstack/openstack Project: openstack-dev/devstack  a8980996340d79bc6fc5a1d5545f3ff6c5709eab
Set config file to db sync and other manage operation

Currently, the db sync operation does not specify the config dir or
config file.
If there is a config file in the home path, it will use this one,
but not the right one devstack write.

Set config file to these operations.

Change-Id: Id1fbc3d85280c19596f5ebd301c46bcf018fa2f6
Closes-Bug: #1578098