Updated openstack/openstack Project: openstack/ironic-python-agent  89f4f33128c7e86e44faf61aa75fe5e2080f8ab9
TinyIPA: Ensure that kernel modules are loaded at boot

Installing packages into a chroot was not something tce-load was ever
designed for, because of this it gets a little confused about loading
kernel modules when we install them. This patch manually calls depmod to
ensure that the kernel modules we install during the build process get
loaded when we boot the ramdisk.

Change-Id: I940d11f3f9f2a39cba16e121dc1809a42690d350