Updated openstack/openstack Project: openstack/requirements  966ce44d95aece2adaf9f878d20f8bd176adcc91
Add pyldap to g-r

pyldap is the python3 port of the python-ldap library. We should add this
as it is a drop-in replacement for python-ldap.

The library is actively maintained by Red Hat and a general community.

The library is PSF licensed, compatible with ASLv2 and OpenStack.

The library is used within Red Hat and therefore has active packaging work.
Other distros should not have a hard time packaging it if it is not included.

This library will replace python-ldap in the long run, so functionality is
duplicated. This library and python-ldap should not be co-installed since they
act in the same namespace. Pyldap is a drop-in replacement and therefore is
API compatible.

This is being added in an effort to get keystone to python 3 compat.

Change-Id: I330b9b8fd3b53a588f0a1ae0c7de16cf03adb3cd