Updated openstack/openstack Project: openstack/requirements  a8a63665cb6badf5d7fceb30215912d8a3c002d5
Add pydotplus for usage during testing

Why: this (optional) library is used to ensure that the pydot
     conversion of taskflow execution workflows into
     dot graphs (which can then be converted into svg or
     png or jpg) works as expected and can be verified during
     taskflow testing (this also applies for consuming projects
     that also want to use/verify pydot, ie consumers of taskflow
     or automaton such as ironic or octavia).

Is the library actively maintained?

  Not as much as I want it to be, but the dot format hasn't changed
  in years and as we encounter bugs I'm ok with trying to work with
  the author to fix them.

Is the library good code?

  Sure?

Is the library python 3 compatible?

  Yes

Is the library license compatible?

  MIT License

Is the library already packaged in the distros
we target (Ubuntu latest / Fedora latest)?

  Seems packaged on fedora, but not ubuntu.

Is the function of this library already covered by other
libraries in global-requirements.txt?

  Nope

Related-Bug: #1561656

Change-Id: Ia633ccd1dca94f70b05ae4376a1c3a3f252a9923