Updated openstack/openstack Project: openstack-infra/project-config  81c5684980687420929308c4f7131d52522e0690
ensure all projects syncing global requirements use check-requirements

Add a job to openstack/requirements to run the new validation command
that ensures that all repos listed in projects.txt have the
check-requirements job defined. The job only runs against the master
branch, for now. As we improve the reliability of the coverage we may
want to extend this to other branches in the future.

Change-Id: I4b0e8c1aca5e492d7f1d4a7f5c9f135569a27e1c
Depends-On: I61658a7265a77e72401978dda558f41b28a81d92
Signed-off-by: Doug Hellmann <doug@doughellmann.com>