Updated openstack/openstack Project: openstack/trove  29c30e522b9d05536cf32a66e9ee1a2ddea3ad22
Fix MariaDB clusters failing with TypeError

This would apply to any type of cluster that uses the galera strategy
while setting the nics on a create call. When we called cast to set()
the object was a list of lists. The set method can not has a list so
this was causesing a unhashable error.

The change is to make the instance_nics a list of strings (what we
originaly expected) to resolve this issue.

Change-Id: I6b04f8b580720e5791e0977a9347d031cdbf9710
Closes-Bug: #1570602