Updated openstack/openstack Project: openstack/heat  869249a49553fb70acad863ccfa1cdf83cec9864
Add --verbose-update-params option to heat-manage

In case the "heat-manage update-params" utility hangs in the middle of
a run, it is useful to know the exact raw_template or resource that
was being processed. Verbose logging of each raw_template or resource
that is being processed will aid in debugging and checking for corrupt
data (if the heat-manage process has to be killed).

Change-Id: Ic552bdaf234c7b9a6a76c345aab816a171b404ab