Updated openstack/openstack Project: openstack/keystone  92837a887cbd5e1bdc7b94388b684a045f226111
Move the oauth1 abstract base class out of core

This patch moves the oauth1 abstract base class out of core and into
backends/base.py

This removes dependencies where backend code references code in the
core. The reasoning being that the core should know about the backend
interface, but the backends should not know anything about the core
(separation of concerns). And part of the risk here is a potential for
circular dependencies.

Partial-Bug: #1563101

Change-Id: I730798c71a6e8be8f32edd5bddb9e86cabd6ff65