Updated openstack/openstack Project: openstack/neutron  6fb1bd44e48c4fd8823a5254156f509b1224d12c
Enable flavor plugin as a default service plugin

This enables the flavor plugin by default. Since it is a
plugin that only controls flavor listing and associations
there isn't much of a downside to always having it enabled.

If another core plugin implements the flavors extension, this
will conflict with it, but we should standardize on a data model
across plugins for flavors so it will be better to encourage them
use the integrated flavors service plugin.

This also enables the 'flavors' extension for the API tests to
exercise the flavors functionality.

Partially-Implements: bp/multi-l3-backends
Change-Id: I6f27eaa02d6248a7f1ae5e6e81d4a60638cc6b7e