Updated openstack/openstack Project: openstack/horizon  0286703c78063afe6a0ce6983e4dadd5c2ca620c
Preview Page: Modal location is off

There is JavaScript in horizon that is forcing a hardcoded inline
style for the top offset of the modal. This is causing shenanigans
on the theme preview page for the lower sections.

There existed a hardcoded CSS class in _debt.scss that was undoing
this JavaScript hardcode, but neither is needed.  Bootstrap should
work properly by itself.

Closes-bug: #1581669

Change-Id: I3bab355be0c53fdd3c6ba8c91846c6c54ebcee4d