Updated openstack/openstack Project: openstack/neutron  b3e7809421d16bc20183bd81d6a3b0c71b0b8953
Pecan: Handle hyphenated collection with shims

A simple replace of hyphens to underscores is needed on a
check to see if a collection needed to be shimmed or not.

This also adds a fake extension and service plugin to test
this.  This should allow for easier testing of other test
cases.

Change-Id: Id2ddd01a4c437f14c67aed37a182871c1fceff13