Updated openstack/openstack Project: openstack/nova  cdfbb9a668fdcf289ffdfa5252d102e9d3e2ec35
pci: create PCI tracker in RT._init_compute_node

The end goal is to have the ResourceTracker's constructor be responsible
for creating the ComputeNode object/DB record and do all setup of
various resource tracking, including PCI device handling, querying the
aggregate and associated resource pool inventory information and
communication with the virt driver to gather host resources.

Before we can do that, however, we need to consolidate the creation of
various tracking entities into the ResourceTracker._init_compute_node()
method. This patch moves the creation of the
nova.pci.manager.PciDevTracker into the _init_compute_node() method and
cleans up a number of unit tests that were manually setting the
ResourceTracker.compute_node attribute to a fixture value (and in so
doing improperly circumventing the initialization process by relying on
the ResourceTracker.update_available_resource() to populate resource
information in that ComputeNode object.

Change-Id: I39a5436ee86e2d9b90f8f3e185353a04b277d434
Related-bug: #1357453
Related-bug: #1357491