Updated openstack/openstack Project: openstack-infra/project-config  4a779af3da6a07abbb5007d8825a035bcc9c7d95
Build also wheel packages for Python 3

Currently, tox gets many prebuilt wheel packages on Python 2 which
make the creation of the Python 2 virtual environments fast. But
Python 3 requires to build much more dependencies and so the creation
of the Python 3 virtual environments is slower.

This change builds also wheel packages on Python 3.

Change-Id: I60c5b2c55eaf3145c3028ede1cf93a76e766b587
Related-Bug: 1582257