Updated openstack/openstack Project: openstack/governance  fc08ddba8ca15db6e0d71c9b55fa600bedad0bf2
Adjust TripleO tags to reflect trailing cycle

We agreed in our recent weekly meeting[1] applying the
new release:cycle-trailing tag to these repos.  Note that we
actually cut these branches for mitaka before the final coordinated
release, but we have a hard-dependency on the puppet openstack
modules, which have applied for the the trailing tag[2], ergo
we must observe the same release cycle despite aiming to release
around the same time as the coordinated release where possible
(which AIUI is also the intention of the puppet community, but we may
need a buffer of the 2-weeks in the event there are packaging or
other delays close to the final release).

The following projects are updated:

tripleo-heat-templates
instack-undercloud
python-tripleoclient
tripleo-common
tripleo-puppet-elements
puppet-tripleo

[1] http://eavesdrop.openstack.org/irclogs/%23openstack-meeting-alt/
    %23openstack-meeting-alt.2016-04-19.log.html
[2] https://review.openstack.org/#/c/297382/

Change-Id: I9eec08fe12147056f51388617c865f27deb56383