Updated openstack/openstack Project: openstack/governance  8ec65923ad06c748c8e7cecae4e431b998787e86
Adjust TripleO tags to use release:cycle-with-intermediary

Changes the remaining TripleO projects that are release:independent to
release:cycle-with-intermediary, with the exception of diskimage-builder
and dib-utils. The release tag for those 2 projects will stay
release:independent for now, pending further discussion.

As part of this change, tripleo-docs is also updated to be tagged
release:none, which is accurate of its current state. In the future,
we may choose to release docs with a given release, but if we do we can
adjust the tag at that time.

Change-Id: I929c2f1a56f4a6c402b7482fdfcfae58e091bbe9